backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
reportlab==4.0.7
pandas==2.1.4
python-decouple==3.8
python-dotenv==1.0.0
psycopg2-binary==2.9.7
Pillow==10.1.0
factory-boy==3.3.0
faker==20.1.0
